feat: handleErrorResponse
- refactor out the `handleMiddlewareError` to a more general `handleErrorResponse` to allow us to extend this to other error types.
