+++ src/lib/errors/UnauthorizedError.ts
@@ -1,0 +1,1 @@
+export class UnauthorizedError extends Error {}
