+++ src/lib/middleware.ts
@@ -1,13 +1,11 @@
 import projectConfig from '@/config/index';
+import { UnauthorizedError } from '@/lib/errors/UnauthorizedError';
 import logger from '@/lib/logger';
 import prisma from '@/lib/prisma';
-import NextResponseErrorBody from '@/types/NextResponseErrorBody';
 import Session from '@/types/Session';
-import { NextRequest, NextResponse } from 'next/server';
+import { NextRequest } from 'next/server';
 
 const authCookieName = projectConfig.auth.cookieName;
-
-export class UnauthorizedError extends Error {}
 
 export async function authMiddleware(request: NextRequest): Promise<Session> {
   const cookie = request.cookies.get(authCookieName);
@@ -27,14 +25,4 @@ export async function authMiddleware(request: NextRequest): Promise<Session> {
   return {
     user,
   };
-}
-
-export function handleMiddlewareError(
-  error: unknown,
-): NextResponse<NextResponseErrorBody> {
-  if (error instanceof UnauthorizedError) {
-    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-  } else {
-    return NextResponse.json({ error: 'Unknown error' }, { status: 500 });
-  }
 }
